+++ src/r_things.h
@@ -38,9 +38,9 @@ struct particle_t
 	WORD	snext;
 };
 
-extern int	NumParticles;
-extern int	ActiveParticles;
-extern int	InactiveParticles;
+extern WORD	NumParticles;
+extern WORD	ActiveParticles;
+extern WORD	InactiveParticles;
 extern particle_t *Particles;
 
 const WORD NO_PARTICLE = 0xffff;
@@ -53,7 +53,7 @@ inline particle_t *NewParticle (void)
 		result = Particles + InactiveParticles;
 		InactiveParticles = result->tnext;
 		result->tnext = ActiveParticles;
-		ActiveParticles = (int)(result - Particles);
+		ActiveParticles = WORD(result - Particles);
 	}
 	return result;
 }
